refactor: project 페이지 상수 폴더 변경
