+++ src/components/project/index.ts
@@ -1,11 +1,4 @@
 export { default as ProjectCard } from './ProjectCard'
 export { default as ProjectMain } from './ProjectMain'
 export { default as ProjectTitle } from './ProjectTitle'
-export {
-  PROJECTS,
-  PROJECT_TAB_ITEMS,
-  STORE_APP_ICON,
-  STORE_PLAY_ICON,
-  projectCoverSrc,
-} from './projectData'
-export type { ProjectCohort, ProjectEntry, ProjectTabId } from './projectData'
+export { PROJECTS, STORE_APP_ICON, STORE_PLAY_ICON, projectCoverSrc } from '@/constants/project'
